+++ config/location.go
@@ -2,8 +2,9 @@ package config
 
 import (
 	"os"
-	"path"
+	"path/filepath"
 
+	"github.com/adrg/xdg"
 	"github.com/mlange-42/tom/util"
 	"gopkg.in/yaml.v3"
 )
@@ -15,12 +16,10 @@ type Location struct {
 }
 
 func LoadLocations() (map[string]Location, error) {
-	dir, err := GetRootDir()
+	path, err := xdg.ConfigFile(locationsFile)
 	if err != nil {
 		return nil, err
 	}
-
-	path := path.Join(dir, locationsFile)
 
 	if !util.FileExists(path) {
 		return map[string]Location{}, nil
@@ -41,16 +40,15 @@ func LoadLocations() (map[string]Location, error) {
 }
 
 func SaveLocations(locations map[string]Location) error {
-	dir, err := GetRootDir()
-	if err != nil {
-		return err
-	}
-	err = CreateDir(dir)
+	path, err := xdg.ConfigFile(locationsFile)
 	if err != nil {
 		return err
 	}
 
-	path := path.Join(dir, locationsFile)
+	err = CreateDir(filepath.Dir(path))
+	if err != nil {
+		return err
+	}
 
 	file, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0600)
 	if err != nil {